Update detfct.fit.mono.R
Some bits in the code were redundant. Also, I think the weak monotonicity check was incorrect. Initially it checked whether any point after the detfct(g0) was smaller than that detfct(g0), which could allow for increases as long as they were not beyond detfct(g0). I updated it to always check with the previous point, but for weak mono it checks =< and for strong mono it checks < .