Create auglagWrapper.R
auglag does not allow for additioanl arguemnts to be passed to the constraint function, beyond the current parameter valeus. Therefore I had to write a wrapper function for auglag wihtin which I define the constraint function. That way, when the constraint function cannot find the additional arguments such as ddfobj, it looks up one scope and finds it, since it is one of the arguments of auglagWrapper.R